Add unstyledList class to Common.module.css and use it in ChooseOrgPage components
